requirements.txt
uvicorn==0.20.0
fastapi==0.89.1
numpy==1.24.1
Pillow==9.3.0
pytesseract==0.3.10
python-multipart
